+++ src/types.d.ts
@@ -24,7 +24,7 @@ export interface ConfigMembers {
     leoGoAnywhereShortcut: boolean;
 
     // statusBarString: string;
-    statusBarColor: string;
+    // statusBarColor: string;
 
     treeInExplorer: boolean;
     showOpenAside: boolean;